// angular-src/src/app/models/cancelqueuelist.ts
export class CancelQueueList {
    cancel_Queue_Id: number;
    customer_Number: string;
    customer_Name: string;
    siteCount: number;
    full_Cancel: string;
    rmR_Reason_Code: string;
    notice_Date: Date;
    effective_Date: Date;
    cancelledRMR: number;
    balance_Of_Contract: number;
    reference: string;
    memo: string;
    address_1: string;
    address_2: string;
    city: string;
    state: string;
    zipCode: string;
    e_Mail: string;
    primaryPhone: string;
    alternatePhone: string;
}